Project/backend/requirements.txt
fastapi
uvicorn
playwright
httpx
ollama
pydantic
python-multipart
pymisp
python-dotenv
beautifulsoup4
python-dateutil
google-auth-oauthlib
google-api-python-client
google-auth
sqlalchemy
